requirements.txt
streamlit
pytesseract
Pillow
requests
python-docx
openpyxl
pdf2image